requirements.txt
numpy~=1.26.4
beautifulsoup4~=4.12.3
#SciPy
#joblib
#threadpoolctl
scikit-learn~=1.4.1.post1
pandas~=2.2.1
#lxml
nltk~=3.8.1
#pillow
tensorflow
transformers~=4.39.2
torch
flax
#jupyterlab
#notebook
#ipywidgets
# run jupyter notebook locally using the following command:
# jupyter notebook

# https://pypi.org/project/google-cloud-aiplatform/
google-cloud-aiplatform
# https://cloud.google.com/python/docs/reference
google-cloud